+++ app/Repositories/FreelancerRepository.php
@@ -18,18 +18,34 @@ class FreelancerRepository implements FreelancerRepositoryInterface {
     }
     public function UpdateFreelancer($id,array $data)
     {   
+        $freelancerExists = DB::table('freelancers')->where('id',$id)->exists();
+        if (!$freelancerExists){
+            return response()->json(['message'=>'Freelancer does not exist'],404);
+        }
         return DB::table('freelancers')->where('id',$id)->update($data);
     }   
     public function DeleteFreelancer($id)
     {
+        $freelancerExists = DB::table('freelancers')->where('id',$id)->exists();
+        if (!$freelancerExists){
+            return response()->json(['message'=>'Freelancer does not exist'],404);
+        }
         return DB::table('freelancers')->where('id',$id)->delete();
     }
     public function GetAllFreelancers()
     {
+        $freelancerExists = DB::table('freelancers')->isEmpty();
+        if (!$freelancerExists){
+            return response()->json(['message'=>'Database is empty'],204);
+        }
         return DB::table('freelancers')->get();
     }
     public function GetFreelancer($id)
     {
+        $freelancerExists = DB::table('freelancers')->where('id',$id)->exists();
+        if (!$freelancerExists){
+            return response()->json(['message'=>'Freelancer does not exist'],404);
+        }
         return DB::table('freelancers')->where('id',$id)->get();
     }
 }
